refactor(StringAddCalculator): move inner class to another class (feedback)
